Re-made docs hierarchy using sphinx-apidoc
This is a bit of a massive change, sorry!
I moved all of the method descriptions to docstrings
in the code, then made it so those get pulled into the
generated docs.

I also merged the intro with the index, so the landing page should
be pretty nice to look at.

Finally, I started working on converting the docstrings to
numpydoc format